{{cookiecutter.project_slug}}/backend/requirements.txt
argon2==0.1.10
argon2-cffi==18.3.0
autopep8==1.4
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.3
django-redis==4.9.0
flake8==3.5.0
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==1.3.6
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
mccabe==0.6.1
mypy==0.620
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.20.0
rope==0.11.0
simplegeneric==0.8.1
six==1.14.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.15.3
{% if cookiecutter.api == 'REST' %}
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
{% elif cookiecutter.api == 'GraphQL' %}
django-graphql-jwt==0.1.10
graphene==2.1.6
graphene-django==2.3.2
graphql-core==2.2
graphql-relay==0.4.5
{% endif %}
{% if cookiecutter.db == 'MongoDB' %}
djongo==1.2.32
{% elif cookiecutter.db == 'PostgreSQL' %}
psycopg2-binary==2.8.4
{% endif %}
